migrated transcript component right inside capture page
