+++ apps/AppWithWearable/lib/main.dart
@@ -96,7 +96,7 @@ class _MyAppState extends State<MyApp> {
           )),
       themeMode: ThemeMode.dark,
       // home: const HasBackupPage(),
-      home: (!SharedPreferencesUtil().onboardingCompleted) //  && SharedPreferencesUtil().deviceId != ''
+      home: (SharedPreferencesUtil().onboardingCompleted) //  && SharedPreferencesUtil().deviceId != ''
           ? const HomePageWrapper()
           : const OnboardingWrapper(),
     );
