+++ apps/AppWithWearable/lib/pages/capture/widgets/widgets.dart
@@ -1,12 +1,19 @@
+import 'dart:convert';
+
 import 'package:flutter/material.dart';
+import 'package:friend_private/backend/mixpanel.dart';
 import 'package:friend_private/backend/preferences.dart';
 import 'package:friend_private/backend/schema/bt_device.dart';
+import 'package:friend_private/backend/storage/segment.dart';
+import 'package:friend_private/pages/speaker_id/page.dart';
 import 'package:friend_private/widgets/device_widget.dart';
 import 'package:friend_private/widgets/scanning_ui.dart';
+import 'package:record/record.dart';
 import 'package:url_launcher/url_launcher.dart';
 
-getConnectedDeviceWidgets(bool hasTranscripts, BTDeviceStruct? device) {
+getConnectionStateWidgets(bool hasTranscripts, BTDeviceStruct? device) {
   if (hasTranscripts) return [];
+    debugPrint('device: $device hasTranscripts: $hasTranscripts');
   if (device == null) {
     return [
       const DeviceAnimationWidget(sizeMultiplier: 0.7),
@@ -132,5 +139,160 @@ _getNoFriendConnectedYet() {
       //   textAlign: TextAlign.center,
       // ),
     ],
+  );
+}
+
+speechProfileWidget(BuildContext context) {
+  return SharedPreferencesUtil().hasSpeakerProfile
+      ? const SizedBox(height: 16)
+      : Stack(
+          children: [
+            GestureDetector(
+              onTap: () {
+                Navigator.of(context).push(MaterialPageRoute(builder: (c) => const SpeakerIdPage()));
+                MixpanelManager().speechProfileCapturePageClicked();
+              },
+              child: Container(
+                decoration: BoxDecoration(
+                  color: Colors.grey.shade900,
+                  borderRadius: const BorderRadius.all(Radius.circular(12)),
+                ),
+                margin: const EdgeInsets.symmetric(vertical: 16, horizontal: 24),
+                padding: const EdgeInsets.all(16),
+                child: const Row(
+                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                  children: [
+                    Expanded(
+                      child: Row(
+                        children: [
+                          Icon(Icons.multitrack_audio),
+                          SizedBox(width: 16),
+                          Text(
+                            'Set up speech profile',
+                            style: TextStyle(color: Colors.white, fontSize: 16),
+                          ),
+                        ],
+                      ),
+                    ),
+                    Icon(Icons.arrow_forward_ios)
+                  ],
+                ),
+              ),
+            ),
+            Positioned(
+              top: 12,
+              right: 24,
+              child: Container(
+                width: 12,
+                height: 12,
+                decoration: const BoxDecoration(color: Colors.red, shape: BoxShape.circle),
+              ),
+            ),
+          ],
+        );
+}
+
+getTranscriptWidget(bool memoryCreating, List<TranscriptSegment> segments, BTDeviceStruct? btDevice) {
+  if (memoryCreating) {
+    return const Padding(
+        padding: EdgeInsets.only(top: 80), child: Center(child: CircularProgressIndicator(color: Colors.white)));
+  }
+
+  if (segments.isEmpty) {
+    return btDevice != null
+        ? const Column(
+            mainAxisSize: MainAxisSize.min,
+            children: [
+              SizedBox(height: 80),
+              Align(
+                alignment: Alignment.center,
+                child: Padding(
+                  padding: EdgeInsets.symmetric(horizontal: 32.0),
+                  child: Text(
+                    textAlign: TextAlign.center,
+                    'Your transcripts will start appearing\nhere after 30 seconds.',
+                    style: TextStyle(color: Colors.white, height: 1.5, decoration: TextDecoration.underline),
+                  ),
+                ),
+              )
+            ],
+          )
+        : const SizedBox.shrink();
+  }
+  // Capture messages
+  var needsUtf8 = SharedPreferencesUtil().recordingsLanguage != 'en';
+  return ListView.separated(
+    padding: EdgeInsets.zero,
+    shrinkWrap: true,
+    scrollDirection: Axis.vertical,
+    itemCount: segments.length + 2,
+    physics: const NeverScrollableScrollPhysics(),
+    separatorBuilder: (_, __) => const SizedBox(height: 16.0),
+    itemBuilder: (context, idx) {
+      if (idx == 0) return const SizedBox(height: 32);
+      if (idx == segments.length + 1) return const SizedBox(height: 64);
+      final data = segments[idx - 1];
+      return Padding(
+        padding: const EdgeInsetsDirectional.fromSTEB(16.0, 0.0, 16.0, 0.0),
+        child: Column(
+          mainAxisSize: MainAxisSize.min,
+          children: [
+            Row(
+              mainAxisAlignment: MainAxisAlignment.start,
+              crossAxisAlignment: CrossAxisAlignment.center,
+              children: [
+                Image.asset(data.isUser ? 'assets/images/speaker_0_icon.png' : 'assets/images/speaker_1_icon.png',
+                    width: 26, height: 26),
+                const SizedBox(width: 12),
+                Text(
+                  data.isUser ? 'You' : 'Speaker ${data.speakerId}',
+                  style: const TextStyle(color: Colors.white, fontSize: 18),
+                )
+              ],
+            ),
+            const SizedBox(height: 12),
+            Align(
+              alignment: Alignment.centerLeft,
+              child: SelectionArea(
+                child: Text(
+                  needsUtf8 ? utf8.decode(data.text.toString().codeUnits) : data.text,
+                  style: const TextStyle(letterSpacing: 0.0, color: Colors.grey),
+                  textAlign: TextAlign.left,
+                ),
+              ),
+            ),
+          ],
+        ),
+      );
+    },
+  );
+}
+
+getPhoneMicRecordingButton(VoidCallback recordingToggled, RecordState state) {
+  if (SharedPreferencesUtil().deviceId.isNotEmpty) return const SizedBox.shrink();
+  return Padding(
+    padding: const EdgeInsets.only(bottom: 140),
+    child: Align(
+      alignment: Alignment.bottomCenter,
+      child: MaterialButton(
+        shape: RoundedRectangleBorder(
+            borderRadius: BorderRadius.circular(12),
+            side: BorderSide(color: state == RecordState.record ? Colors.red : Colors.white)),
+        onPressed: recordingToggled,
+        child: Padding(
+          padding: const EdgeInsets.symmetric(horizontal: 8, vertical: 12),
+          child: Row(
+            mainAxisSize: MainAxisSize.min,
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: [
+              state == RecordState.record ? const Icon(Icons.stop, color: Colors.red, size: 24) : const Icon(Icons.mic),
+              const SizedBox(width: 8),
+              Text(state == RecordState.record ? 'Stop Recording' : 'Try With Phone Mic'),
+              const SizedBox(width: 4),
+            ],
+          ),
+        ),
+      ),
+    ),
   );
 }
